rename body -> range
