+++ src/components/Scale/range.tsx
