+++ src/index.tsx
@@ -30,7 +30,7 @@ import * as _DatePicker from './components/Datepicker/index'
 import * as _DateRangePicker from './components/DateRangePicker/index'
 
 import * as _ScaleController from './components/Scale/controller'
-import * as _ScaleBody from './components/Scale/body'
+import * as _ScaleRange from './components/Scale/range'
 
 export const Text = _Text.Component
 export const Icon = _Icon.Component
@@ -58,4 +58,4 @@ export const DatePicker = _DatePicker.Component
 export const DateRangePicker = _DateRangePicker.Component
 export const Modal = _Modal.Component
 export const ScaleController = _ScaleController.Component
-export const ScaleBody = _ScaleBody.Component
+export const ScaleRange = _ScaleRange.Component
